# docs/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: AgentSwarm Task Pool
  version: 0.5.0
  description: AgentSwarm pull-based task pool (Phases 0–4, credibility gamification, memory gating)
servers:
  - url: http://localhost:8000
paths:
...
  /tasks:
    post:
      summary: Create a task (owner auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
      responses:
        "200":
          description: Created
...
  /tasks/{task_id}:
    get:
      summary: Get task by id
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task envelope
...
components:
  schemas:
...
    TaskCreateRequest:
      type: object
      required: [task_type, capability_required]
      properties:
        task_type:
          type: string
        capability_required:
          type: string
        payload:
          type: object
          description: May include stake_tier (low, medium, high)
        project_id:
          type: string
        parent_task_id:
          type: string
        parent_submission_id:
          type: string
...